Switch to all fixed point calculations inside virtual machine
This increases the performance of the virtual machine.

All dimensions both internal and passed in / out of gvm area now long int in
micro metres.

+++ point.h
@@ -11,11 +11,14 @@ enum axismask {
 };
 
 
+/*
+ * Dimensions is units of micro meters.
+ */
 struct point {
-	float x;
-	float y;
-	float z;
-	float e;
+	long int x;
+	long int y;
+	long int z;
+	long int e;
 };
